マイグレーションファイル作り直し
