--- app/controllers/categories_controller.rb
@@ -1,5 +1,0 @@
-class CategoriesController < ApplicationController
-  def index
-    @categories = Category.all
-  end
-end
