--- app/helpers/categories_helper.rb
@@ -1,2 +1,0 @@
-module CategoriesHelper
-end
