--- app/models/category.rb
@@ -1,3 +1,0 @@
-class Category < ApplicationRecord
-  belongs_to :event
-end
