+++ app/views/events/new.html.erb
@@ -30,10 +30,9 @@
     <%= form.text_field :location %>
   </div>
 
-
     <div class="category">
     <%# form.label :category_id, "カテゴリー" %>
-    <%# form.collection_select(:category_id, [], :id, :name, { prompt: "カテゴリーを選択してください" }, {class:"select-box", id:"item-sales-status"}) %>
+    <%# form.select :category, ["Work", "Friend", "Themepark", "Hobby", "Other", "period"], :prompt => "選択してください" %>
   </div>
 
   <div class="actions">
