--- db/migrate/20240227073436_add_user_id_to_events.rb
@@ -1,5 +1,0 @@
-class AddUserIdToEvents < ActiveRecord::Migration[7.0]
-  def change
-    add_reference :events, :user, null: false, foreign_key: true
-  end
-end
