+++ db/migrate/20240308140918_remove_category_id_from_events.rb
@@ -1,0 +1,5 @@
+class RemoveCategoryIdFromEvents < ActiveRecord::Migration[7.0]
+  def change
+    remove_column :events, :category_id
+  end
+end
