+++ db/migrate/20240308143334_remove_user_id_from_events.rb
@@ -1,0 +1,5 @@
+class RemoveUserIdFromEvents < ActiveRecord::Migration[7.0]
+  def change
+    remove_column :events, :user_id
+  end
+end
