+++ db/migrate/20240308143601_add_user_id_to_events.rb
@@ -1,0 +1,5 @@
+class AddUserIdToEvents < ActiveRecord::Migration[7.0]
+  def change
+    add_reference :events, :user, null: false, foreign_key: true unless column_exists?(:events, :user_id)
+  end
+end
