--- test/controllers/categories_controller_test.rb
@@ -1,7 +1,0 @@
-require "test_helper"
-
-class CategoriesControllerTest < ActionDispatch::IntegrationTest
-  # test "the truth" do
-  #   assert true
-  # end
-end
